+++ apps/notification-center/CHANGELOG.md
@@ -1,5 +1,29 @@
 # notification-center
 
+## 1.0.35
+
+### Patch Changes
+
+- d674c81: Update dependencies
+- Updated dependencies [d674c81]
+  - @mgdis/notification-center@1.1.10
+  - @mgdis/mg-components@5.27.2
+
+## 1.0.34
+
+### Patch Changes
+
+- Updated dependencies [17399c8]
+- Updated dependencies [3f8fbfa]
+- Updated dependencies [f228d36]
+- Updated dependencies [0c93b4e]
+- Updated dependencies [28b0455]
+- Updated dependencies [140a4ab]
+- Updated dependencies [f1f76dc]
+- Updated dependencies [3d998cc]
+  - @mgdis/mg-components@5.27.1
+  - @mgdis/notification-center@1.1.9
+
 ## 1.0.33
 
 ### Patch Changes