+++ configs/eslint-config-stencil/CHANGELOG.md
@@ -1,5 +1,21 @@
 # eslint-config-stencil
 
+## 1.1.9
+
+### Patch Changes
+
+- d674c81: Update dependencies
+- Updated dependencies [d674c81]
+  - eslint-config-custom@1.0.8
+
+## 1.1.8
+
+### Patch Changes
+
+- 140a4ab: Update dependencies
+- Updated dependencies [140a4ab]
+  - eslint-config-custom@1.0.7
+
 ## 1.1.7
 
 ### Patch Changes