+++ packages/linting-stencil/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @mgdis/linting-stencil
 
+## 1.1.4
+
+### Patch Changes
+
+- d674c81: Update dependencies
+
+## 1.1.3
+
+### Patch Changes
+
+- 140a4ab: Update dependencies
+
 ## 1.1.2
 
 ### Patch Changes