+++ packages/mg-components-helpers/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 4.1.2
+
+### Patch Changes
+
+- d674c81: Update dependencies
+
+## 4.1.1
+
+### Patch Changes
+
+- 140a4ab: Update dependencies
+
 ## 4.1.0
 
 ### Minor Changes