+++ packages/styles/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @mgdis/styles
 
+## 1.3.2
+
+### Patch Changes
+
+- d674c81: Update dependencies
+
+## 1.3.1
+
+### Patch Changes
+
+- 140a4ab: Update dependencies
+
 ## 1.3.0
 
 ### Minor Changes